refactor homepage
