+++ .prettierrc.json
@@ -3,6 +3,5 @@
   "tabWidth": 2,
   "semi": false,
   "singleQuote": true,
-  "plugins": ["prettier-plugin-tailwindcss"],
-  "pluginSearchDirs": false
+  "plugins": ["prettier-plugin-tailwindcss"]  
 }
