--- app/(site)/portableText.tsx
@@ -1,17 +1,0 @@
-import { PortableTextComponents } from '@portabletext/react'
-
-export const homePortableText: PortableTextComponents = {
-  marks: {
-    strong: ({ children }) => (
-      <strong className=" dark:text-purple-dark">{children}</strong>
-    ),
-  },
-  block: {
-    normal: ({ children }) => <p className="mt-4 md:mr-4">{children}</p>,
-    h1: ({ children }) => (
-      <h1 className="text-2xl text-purple-light dark:text-purple-dark md:mt-0 md:text-4xl">
-        {children}
-      </h1>
-    ),
-  },
-}
