+++ components/Analytics.tsx
