+++ components/FeaturedProjects/FeaturedProjects.tsx
