+++ components/Icons.tsx
@@ -127,19 +127,6 @@ export function OpenOutline({ size }: Icon) {
   )
 }
 
-export function TriangleSharp({ size }: Icon) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 512 512"
-    >
-      <path fill="currentColor" d="M256 32L20 464h472L256 32z" />
-    </svg>
-  )
-}
-
 export function Moon({ size = 12 }: Icon) {
   return (
     <svg
@@ -244,20 +231,6 @@ export function NEXTJS({ size = 124, classProps }: Icon) {
   )
 }
 
-export function JS({ size = 32, classProps }: Icon) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 256 256"
-    >
-      <path fill="#F7DF1E" d="M0 0h256v256H0V0Z" />
-      <path d="m67.312 213.932l19.59-11.856c3.78 6.701 7.218 12.371 15.465 12.371c7.905 0 12.89-3.092 12.89-15.12v-81.798h24.057v82.138c0 24.917-14.606 36.259-35.916 36.259c-19.245 0-30.416-9.967-36.087-21.996m85.07-2.576l19.588-11.341c5.157 8.421 11.859 14.607 23.715 14.607c9.969 0 16.325-4.984 16.325-11.858c0-8.248-6.53-11.17-17.528-15.98l-6.013-2.58c-17.357-7.387-28.87-16.667-28.87-36.257c0-18.044 13.747-31.792 35.228-31.792c15.294 0 26.292 5.328 34.196 19.247l-18.732 12.03c-4.125-7.389-8.591-10.31-15.465-10.31c-7.046 0-11.514 4.468-11.514 10.31c0 7.217 4.468 10.14 14.778 14.608l6.014 2.577c20.45 8.765 31.963 17.7 31.963 37.804c0 21.654-17.012 33.51-39.867 33.51c-22.339 0-36.774-10.654-43.819-24.574" />
-    </svg>
-  )
-}
-
 export function TS({ size = 32, classProps }: Icon) {
   return (
     <svg
@@ -294,48 +267,6 @@ export function MDB({ size = 121.37, classProps }: Icon) {
         fill="#001E2B"
         d="m125.59 95.636l-21.53-52.583l-.047-.145h-16.75v3.534h2.702c.821 0 1.593.34 2.173.92c.579.581.869 1.356.869 2.18l-.483 52.92c0 1.647-1.351 3.002-2.993 3.051l-2.751.048v3.486h16.316v-3.486l-1.69-.048c-1.641-.049-2.993-1.404-2.993-3.05V52.59l23.46 56.456c.338.823 1.11 1.356 1.98 1.356c.868 0 1.64-.533 1.978-1.356l22.93-55.197l.338 48.613c0 1.694-1.352 3.05-3.042 3.098h-1.738v3.486h19.116v-3.486h-2.607c-1.64 0-2.993-1.404-3.04-3.05l-.146-52.922a3.097 3.097 0 0 1 2.993-3.098l2.8-.049v-3.534h-16.316l-21.53 52.728Zm150.113 9.004c-.532-.533-.823-1.26-.823-2.134V76.51c0-4.947-1.451-8.828-4.355-11.592c-2.856-2.765-6.824-4.171-11.76-4.171c-6.922 0-12.39 2.813-16.214 8.342c-.048.097-.193.146-.338.146c-.146 0-.242-.098-.242-.243l-1.791-6.936h-3.001l-7.695 4.414v2.425h1.984c.92 0 1.694.242 2.227.728c.532.485.822 1.212.822 2.23v30.605c0 .873-.29 1.601-.822 2.135c-.533.533-1.259.824-2.13.824h-1.936v3.54h17.713v-3.54h-1.935c-.872 0-1.598-.291-2.13-.824c-.532-.534-.823-1.262-.823-2.135V82.184c0-2.57.581-5.141 1.646-7.664c1.113-2.473 2.758-4.559 4.936-6.16c2.178-1.6 4.792-2.376 7.792-2.376c3.388 0 5.953 1.067 7.55 3.201c1.597 2.134 2.42 4.899 2.42 8.197v25.027c0 .873-.29 1.6-.823 2.134c-.532.534-1.258.824-2.13.824h-1.935v3.541h17.713v-3.54h-1.935c-.726.097-1.404-.194-1.985-.728Zm161.833-57.773c-4.896-2.607-10.368-3.96-16.273-3.96h-23.04v3.525h2.255c.864 0 1.632.338 2.4 1.11c.72.725 1.104 1.546 1.104 2.415v51.95c0 .87-.383 1.69-1.104 2.415c-.72.724-1.536 1.11-2.4 1.11h-2.256v3.525h23.041c5.905 0 11.377-1.352 16.273-3.96c4.896-2.607 8.88-6.47 11.76-11.394c2.881-4.924 4.369-10.863 4.369-17.623c0-6.759-1.488-12.65-4.368-17.622c-2.928-5.021-6.865-8.836-11.76-11.491Zm6.865 29.017c0 6.18-1.104 11.394-3.265 15.595c-2.16 4.2-5.04 7.338-8.592 9.367c-3.552 2.027-7.489 3.041-11.713 3.041h-4.656c-.864 0-1.632-.338-2.4-1.11c-.72-.725-1.104-1.545-1.104-2.414V51.357c0-.87.336-1.641 1.104-2.414c.72-.724 1.536-1.11 2.4-1.11h4.656c4.224 0 8.16 1.013 11.713 3.041c3.552 2.028 6.432 5.166 8.592 9.367c2.16 4.248 3.265 9.511 3.265 15.643Zm63.728 3.525c-2.131-2.463-6.244-4.539-11.08-5.65c6.678-3.331 10.11-8.014 10.11-14.05c0-3.282-.867-6.228-2.612-8.738c-1.74-2.51-4.21-4.539-7.353-5.939c-3.148-1.4-6.826-2.124-10.987-2.124h-26.081v3.524h2.08c.871 0 1.646.338 2.42 1.11c.726.725 1.113 1.546 1.113 2.415v51.95c0 .87-.387 1.69-1.113 2.415c-.726.724-1.549 1.11-2.42 1.11h-2.274v3.525h28.31c4.305 0 8.32-.724 11.951-2.173c3.63-1.448 6.533-3.573 8.615-6.373c2.127-2.8 3.192-6.229 3.192-10.188c-.048-4.248-1.307-7.87-3.871-10.814Zm-32.47 23.416c-.726-.724-1.113-1.545-1.113-2.414V77.09H488c4.739 0 8.369 1.207 10.886 3.62c2.516 2.415 3.775 5.553 3.775 9.416c0 2.317-.584 4.587-1.644 6.663c-1.114 2.124-2.763 3.814-4.985 5.118c-2.18 1.303-4.889 1.98-8.032 1.98h-9.923c-.87 0-1.645-.339-2.419-1.063Zm-1.064-30.707V51.405c0-.869.338-1.641 1.112-2.414c.726-.724 1.549-1.11 2.42-1.11h6.388c4.594 0 7.984 1.159 10.115 3.38c2.126 2.269 3.191 5.166 3.191 8.738c0 3.67-1.017 6.615-2.999 8.836c-1.986 2.172-4.985 3.283-8.952 3.283h-11.275ZM208.27 63.825c-3.698-2.02-7.828-3.08-12.294-3.08c-4.348 0-8.423.959-12.004 2.92l-.289.16c-3.698 2.021-6.627 4.956-8.788 8.661c-2.16 3.705-3.265 8.036-3.265 12.847c0 4.812 1.104 9.143 3.265 12.848s5.09 6.64 8.788 8.66c3.698 2.021 7.828 3.08 12.293 3.08c4.349 0 8.424-.958 12.004-2.918l.29-.161c3.697-2.021 6.627-4.956 8.788-8.661c2.16-3.705 3.265-8.036 3.265-12.848c0-4.811-1.105-9.142-3.265-12.847c-2.161-3.705-5.09-6.64-8.788-8.66Zm3.553 21.508c0 5.919-1.44 10.73-4.322 14.195c-2.833 3.464-6.722 5.244-11.525 5.244c-4.716 0-8.552-1.717-11.372-5.06l-.152-.184c-2.882-3.465-4.323-8.276-4.323-14.195c0-5.81 1.39-10.554 4.167-14.004l.156-.19c2.832-3.464 6.723-5.245 11.524-5.245c4.803 0 8.692 1.78 11.525 5.245c2.882 3.465 4.322 8.276 4.322 14.194Zm166.631-21.508c-3.698-2.02-7.827-3.08-12.293-3.08c-4.348 0-8.424.959-12.004 2.92l-.29.16c-3.697 2.021-6.626 4.956-8.787 8.661c-2.161 3.705-3.266 8.036-3.266 12.847c0 4.812 1.105 9.143 3.266 12.848c2.16 3.705 5.09 6.64 8.788 8.66c3.697 2.021 7.827 3.08 12.293 3.08c4.348 0 8.424-.958 12.004-2.918l.29-.161c3.697-2.021 6.626-4.956 8.787-8.661c2.161-3.705 3.265-8.036 3.265-12.848c0-4.811-1.104-9.142-3.265-12.847c-2.16-3.705-5.138-6.64-8.788-8.66Zm3.554 21.508c0 5.919-1.441 10.73-4.322 14.195c-2.834 3.464-6.723 5.244-11.525 5.244c-4.717 0-8.553-1.717-11.372-5.06l-.153-.184c-2.881-3.465-4.322-8.276-4.322-14.195c0-5.858 1.389-10.556 4.166-14.004l.156-.19c2.833-3.464 6.723-5.245 11.525-5.245c4.802 0 8.691 1.78 11.525 5.245c2.833 3.465 4.322 8.276 4.322 14.194Zm-72.186-24.587c-3.87 0-7.4.821-10.592 2.465c-3.192 1.643-5.707 3.867-7.496 6.719a17.043 17.043 0 0 0-2.708 9.281c0 2.997.677 5.752 2.08 8.218c1.353 2.368 3.191 4.35 5.513 5.993l-6.916 9.378c-.87 1.16-.967 2.707-.339 3.964c.677 1.305 1.935 2.078 3.386 2.078h1.983c-1.854 1.251-3.353 2.724-4.411 4.46l-.136.229c-1.257 2.078-1.886 4.254-1.886 6.477c0 4.083 1.773 7.467 5.272 10.013l.193.138c3.58 2.562 8.609 3.867 14.944 3.867c4.401 0 8.609-.725 12.43-2.127c3.758-1.361 6.831-3.362 9.132-5.956l.2-.23c2.37-2.708 3.58-5.995 3.58-9.765c0-3.964-1.451-6.768-4.836-9.474c-2.902-2.272-7.448-3.48-13.107-3.48h-19.344a.139.139 0 0 1-.077-.032l-.02-.017l-.009-.02c-.01-.031-.026-.09.009-.125l5.03-6.767a17.789 17.789 0 0 0 3.724 1.256c1.16.242 2.466.339 3.917.339c4.062 0 7.738-.822 10.93-2.465c3.192-1.644 5.755-3.867 7.593-6.72c1.837-2.803 2.756-5.945 2.756-9.28c0-3.577-1.74-10.103-6.48-13.438c0-.03.017-.041.031-.046l.017-.003l10.398 1.16v-4.785h-16.637c-2.611-.822-5.32-1.305-8.124-1.305Zm5.803 30.26c-1.838.966-3.82 1.498-5.803 1.498c-3.24 0-6.094-1.16-8.512-3.432c-2.418-2.272-3.627-5.607-3.627-9.861c0-4.254 1.209-7.59 3.627-9.861c2.418-2.272 5.271-3.432 8.512-3.432c1.946 0 3.804.444 5.573 1.374l.23.124c1.838.967 3.337 2.466 4.546 4.448c1.16 1.982 1.79 4.447 1.79 7.347c0 2.949-.581 5.414-1.79 7.347c-1.16 1.982-2.708 3.48-4.546 4.447Zm-13.106 17.788h13.106c3.627 0 5.949.725 7.496 2.272c1.548 1.547 2.321 3.625 2.321 6.042c0 3.529-1.402 6.43-4.207 8.604c-2.805 2.176-6.577 3.287-11.22 3.287c-4.063 0-7.448-.918-9.915-2.658c-2.466-1.74-3.723-4.4-3.723-7.783c0-2.127.58-4.109 1.741-5.849c1.16-1.74 2.563-2.997 4.4-3.915Z"
       />
-    </svg>
-  )
-}
-
-export function ASTRO({ size = 31.04, classProps }: Icon) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox="0 0 256 264"
-    >
-      <defs>
-        <linearGradient
-          id="logosAstro0"
-          x1="137.073%"
-          x2="85.999%"
-          y1="-40.996%"
-          y2="56.205%"
-        >
-          <stop offset="0%" stopColor="#FF1639" />
-          <stop offset="100%" stopColor="#FF1639" stopOpacity="0" />
-        </linearGradient>
-      </defs>
-      <g fill="none" fillRule="evenodd">
-        <path
-          fill="#000"
-          d="M156.377 4.834c1.575 1.956 2.379 4.595 3.986 9.874l35.105 115.32a146.081 146.081 0 0 0-41.97-14.209l-22.857-77.24a2.975 2.975 0 0 0-5.708.008l-22.58 77.193a146.086 146.086 0 0 0-42.159 14.23L95.471 14.682c1.612-5.27 2.418-7.904 3.994-9.857A12.97 12.97 0 0 1 104.72.936C107.048 0 109.804 0 115.315 0h25.196c5.518 0 8.277 0 10.607.938a13.016 13.016 0 0 1 5.259 3.896Z"
-        />
-        <path
-          fill="#FF5D01"
-          d="M160.577 135.516c-5.787 4.949-17.338 8.324-30.643 8.324c-16.33 0-30.017-5.084-33.65-11.922c-1.298 3.92-1.589 8.404-1.589 11.269c0 0-.856 14.068 8.929 23.852a9.199 9.199 0 0 1 9.2-9.198c8.707 0 8.697 7.596 8.69 13.76l-.001.55c0 9.355 5.718 17.375 13.85 20.757a18.86 18.86 0 0 1-1.896-8.268c0-8.923 5.238-12.246 11.326-16.108c4.845-3.072 10.227-6.485 13.936-13.332a25.145 25.145 0 0 0 3.034-12.013a25.24 25.24 0 0 0-1.186-7.67Z"
-        />
-        <path
-          fill="url(#logosAstro0)"
-          d="M160.577 135.516c-5.787 4.949-17.338 8.324-30.643 8.324c-16.33 0-30.017-5.084-33.65-11.922c-1.298 3.92-1.589 8.404-1.589 11.269c0 0-.856 14.068 8.929 23.852a9.199 9.199 0 0 1 9.2-9.198c8.707 0 8.697 7.596 8.69 13.76l-.001.55c0 9.355 5.718 17.375 13.85 20.757a18.86 18.86 0 0 1-1.896-8.268c0-8.923 5.238-12.246 11.326-16.108c4.845-3.072 10.227-6.485 13.936-13.332a25.145 25.145 0 0 0 3.034-12.013a25.24 25.24 0 0 0-1.186-7.67Z"
-        />
-        <path
-          fill="#000"
-          d="M28.014 221.64c15.874 0 24.2 5.68 24.2 16.418v10.464c.006 4.954.09 7.704.389 10.311l.062.52c.109.866.242 1.732.405 2.673h-7.704c-.389-2.49-.544-5.214-.544-8.093c-4.591 6.458-12.45 9.571-25.446 9.571c-11.75 0-19.376-4.28-19.376-11.672c0-3.502 1.478-6.537 4.124-8.482c2.802-2.023 6.459-3.424 16.341-3.969l24.279-1.4v-.779c0-6.225-6.225-9.882-16.73-9.882c-10.661 0-17.12 3.035-18.365 8.793H1.48c1.245-8.56 11.205-14.474 26.535-14.474Zm58.86 0c13.384 0 22.021 3.579 25.912 9.649l-6.303 3.034c-3.735-4.824-9.416-7.003-20.31-7.003c-9.571 0-14.007 2.101-14.007 6.147c0 3.19 2.88 4.67 8.015 5.058l18.91 1.479c10.116.856 14.084 4.124 14.084 10.505c0 7.704-6.459 12.995-22.566 12.995c-14.24 0-22.567-4.046-27.314-11.205l6.148-3.346c3.502 5.68 10.738 8.793 22.177 8.793c9.416 0 14.163-2.18 14.163-6.77c0-3.19-1.868-4.436-8.015-4.903l-18.754-1.478c-9.26-.778-14.162-4.046-14.162-10.816c0-7.548 8.248-12.14 22.022-12.14Zm141.19 0c16.575 0 27.936 8.403 27.936 20.932c0 12.528-11.361 20.932-27.936 20.932c-16.574 0-27.935-8.404-27.935-20.932c0-12.529 11.36-20.933 27.935-20.933Zm-91.497-11.673v13.15h19.454v5.837h-19.454v18.053c0 6.692 3.658 9.338 10.427 9.338c3.19 0 6.46-.467 9.183-1.09v6.304c-1.946.856-6.381 1.556-10.038 1.556c-14.396 0-17.042-7.704-17.042-15.407v-18.754H116.88v-5.836h12.217v-10.116l7.47-3.035Zm55.203 11.984c2.1 0 3.502.077 5.136.467v7.159c-1.79-.312-3.113-.467-5.214-.467c-6.147 0-9.649 1.712-12.373 4.746c-3.112 3.269-4.202 8.015-4.202 13.618v14.552h-7.392v-38.908h6.848V234.4c2.256-7.548 8.87-12.45 17.197-12.45ZM44.744 243.039l-22.8 1.4c-10.972.7-14.085 2.724-14.085 7.237c0 4.202 4.98 6.303 14.007 6.303c13.307 0 22.878-4.98 22.878-13.15v-1.79Zm183.32-15.252c-12.217 0-20.076 5.758-20.076 14.785c0 9.026 7.86 14.785 20.076 14.785c12.14 0 20.077-5.759 20.077-14.785c0-9.027-7.938-14.785-20.077-14.785Z"
-        />
-      </g>
     </svg>
   )
 }
